优化灰度发布管理平台核心逻辑

+++ book-gray-server/src/main/java/cn/dingyuegroup/gray/server/mysql/entity/GrayPolicyEntity.java
@@ -14,7 +14,7 @@ public class GrayPolicyEntity {
 
     private String policy;
 
-    private Short isDelete;
+    private Integer isDelete;
 
     private Date createTime;
