+++ book-gray-server/src/main/java/cn/dingyuegroup/gray/server/mysql/entity/GrayPolicyGroupEntity.java
@@ -13,9 +13,9 @@ public class GrayPolicyGroupEntity {
 
     private String alias;
 
-    private Short enable;
+    private Integer enable;
 
-    private Short isDelete;
+    private Integer isDelete;
 
     private List<GrayPolicyEntity> grayPolicyEntities = new ArrayList<>();
 }
